+++ hamonize-center/src/main/java/com/controller/curl/CurlAgentUpdtController.java
@@ -46,7 +46,7 @@ public class CurlAgentUpdtController {
 		uuid = uuid.trim();
 
 		// uuid로 부서정보 가져오기
-		int segSeq = pcUUID(uuid);
+		Long segSeq = pcUUID(uuid);
 		if (segSeq == 0) {
 			return "nodata";
 		}
@@ -218,12 +218,12 @@ public class CurlAgentUpdtController {
 	 * 
 	 * @return 부서seq
 	 */
-	public int pcUUID(String uuid) {
+	public Long pcUUID(String uuid) {
 		GetAgentJobVo agentVo = new GetAgentJobVo();
 		agentVo.setPc_uuid(uuid);
 		agentVo = agentJobMapper.getAgentJobPcUUID(agentVo);
 
-		int segSeq = 0;
+		Long segSeq = 0L;
 		if (agentVo != null) {
 			segSeq = agentVo.getSeq();
 		}
